fix: rename folder to original name
