+++ src/components/formAtleta/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 import styles from "./formAtleta.module.css";
 import Loading from "@/components/loading/index";
-import FormTitle from "@/components/titlee/formTitle/index";
+import FormTitle from "@/components/title/formTitle/index";
 import { useForm } from "react-hook-form";
 import { useState } from "react";
 import { TAtleta } from "@/types/TAtleta";
